+++ ClickHaskell.hs
@@ -103,7 +103,6 @@ import Data.Text.Encoding as Text (encodeUtf8)
 import Data.Time (UTCTime, ZonedTime, zonedTimeToUTC)
 import Data.Time.Clock.POSIX (posixSecondsToUTCTime, utcTimeToPOSIXSeconds)
 import Data.Typeable (Proxy (..))
-import Data.Vector.Primitive.Mutable (Prim)
 import Data.Version (Version (..))
 import Data.Word (Word16, Word32, Word8)
 import Debug.Trace (traceShowId)
@@ -1787,7 +1786,7 @@ instance
 
 -- | ClickHouse UUID column type
 newtype ChUUID = MkChUUID Word128
-  deriving newtype (Generic, Show, Eq, NFData, Bounded, Prim, Enum)
+  deriving newtype (Generic, Show, Eq, NFData, Bounded, Enum)
 
 instance IsChType ChUUID where
   type ToChTypeName ChUUID = "UUID"
@@ -1847,7 +1846,7 @@ instance
 
 -- | ClickHouse Int8 column type
 newtype ChInt8 = MkChInt8 Int8
-  deriving newtype (Show, Eq, Num, Prim, Bits, Enum, Ord, Real, Integral, Bounded, NFData)
+  deriving newtype (Show, Eq, Num, Bits, Enum, Ord, Real, Integral, Bounded, NFData)
 
 instance IsChType ChInt8 where
   type ToChTypeName ChInt8 = "Int8"
@@ -1866,7 +1865,7 @@ instance FromChType ChInt8 Int8   where fromChType = coerce
 
 -- | ClickHouse Int16 column type
 newtype ChInt16 = MkChInt16 Int16
-  deriving newtype (Show, Eq, Num, Prim, Bits, Enum, Ord, Real, Integral, Bounded, NFData)
+  deriving newtype (Show, Eq, Num, Bits, Enum, Ord, Real, Integral, Bounded, NFData)
 
 instance IsChType ChInt16 where
   type ToChTypeName ChInt16 = "Int16"
@@ -1883,7 +1882,7 @@ instance FromChType ChInt16 Int16   where fromChType (MkChInt16 int16) = int16
 
 -- | ClickHouse Int32 column type
 newtype ChInt32 = MkChInt32 Int32
-  deriving newtype (Show, Eq, Num, Prim, Bits, Enum, Ord, Real, Integral, Bounded, NFData)
+  deriving newtype (Show, Eq, Num, Bits, Enum, Ord, Real, Integral, Bounded, NFData)
 
 instance IsChType ChInt32 where
   type ToChTypeName ChInt32 = "Int32"
@@ -1900,7 +1899,7 @@ instance FromChType ChInt32 Int32   where fromChType (MkChInt32 int32) = int32
 
 -- | ClickHouse Int64 column type
 newtype ChInt64 = MkChInt64 Int64
-  deriving newtype (Show, Eq, Num, Prim, Bits, Enum, Ord, Real, Integral, Bounded, NFData)
+  deriving newtype (Show, Eq, Num, Bits, Enum, Ord, Real, Integral, Bounded, NFData)
 
 instance IsChType ChInt64 where
   type ToChTypeName ChInt64 = "Int64"
@@ -1918,7 +1917,7 @@ instance FromChType ChInt64 Int64   where fromChType = coerce
 
 -- | ClickHouse Int128 column type
 newtype ChInt128 = MkChInt128 Int128
-  deriving newtype (Show, Eq, Num, Prim, Bits, Ord, Real, Enum, Integral, Bounded, NFData)
+  deriving newtype (Show, Eq, Num, Bits, Ord, Real, Enum, Integral, Bounded, NFData)
 
 instance IsChType ChInt128 where
   type ToChTypeName ChInt128 = "Int128"
@@ -1935,7 +1934,7 @@ instance FromChType ChInt128 Int128   where fromChType (MkChInt128 int128) = int
 
 -- | ClickHouse UInt8 column type
 newtype ChUInt8 = MkChUInt8 Word8
-  deriving newtype (Show, Eq, Num, Prim, Bits, Enum, Ord, Real, Integral, Bounded, NFData)
+  deriving newtype (Show, Eq, Num, Bits, Enum, Ord, Real, Integral, Bounded, NFData)
 
 instance IsChType ChUInt8 where
   type ToChTypeName ChUInt8 = "UInt8"
@@ -1953,7 +1952,7 @@ instance FromChType ChUInt8 Word8   where fromChType (MkChUInt8 w8) = w8
 
 -- | ClickHouse UInt16 column type
 newtype ChUInt16 = MkChUInt16 Word16
-  deriving newtype (Show, Eq, Num, Prim, Bits, Enum, Ord, Real, Integral, Bounded, NFData)
+  deriving newtype (Show, Eq, Num, Bits, Enum, Ord, Real, Integral, Bounded, NFData)
 
 instance IsChType ChUInt16 where
   type ToChTypeName ChUInt16 = "UInt16"
@@ -1970,7 +1969,7 @@ instance FromChType ChUInt16 Word16   where fromChType = coerce
 
 -- | ClickHouse UInt32 column type
 newtype ChUInt32 = MkChUInt32 Word32
-  deriving newtype (Show, Eq, Num, Prim, Bits, Enum, Ord, Real, Integral, Bounded, NFData)
+  deriving newtype (Show, Eq, Num, Bits, Enum, Ord, Real, Integral, Bounded, NFData)
 
 instance IsChType ChUInt32 where
   type ToChTypeName ChUInt32 = "UInt32"
@@ -1987,7 +1986,7 @@ instance FromChType ChUInt32 Word32   where fromChType (MkChUInt32 word32) = wor
 
 -- | ClickHouse UInt64 column type
 newtype ChUInt64 = MkChUInt64 Word64
-  deriving newtype (Show, Eq, Num, Prim, Bits, Enum, Ord, Real, Integral, Bounded, NFData)
+  deriving newtype (Show, Eq, Num, Bits, Enum, Ord, Real, Integral, Bounded, NFData)
 
 instance IsChType ChUInt64 where
   type ToChTypeName ChUInt64 = "UInt64"
@@ -2004,7 +2003,7 @@ instance FromChType ChUInt64 Word64   where fromChType (MkChUInt64 w64) = w64
 
 -- | ClickHouse UInt128 column type
 newtype ChUInt128 = MkChUInt128 Word128
-  deriving newtype (Show, Eq, Num, Prim, Bits, Enum, Ord, Real, Integral, Bounded, NFData)
+  deriving newtype (Show, Eq, Num, Bits, Enum, Ord, Real, Integral, Bounded, NFData)
 
 instance IsChType ChUInt128 where
   type ToChTypeName ChUInt128 = "UInt128"
@@ -2029,7 +2028,7 @@ ClickHouse DateTime column type (paramtrized with timezone)
 "DateTime('UTC')"
 -}
 newtype ChDateTime (tz :: Symbol) = MkChDateTime Word32
-  deriving newtype (Show, Eq, Prim, Num, Bits, Enum, Ord, Real, Integral, Bounded, NFData)
+  deriving newtype (Show, Eq, Num, Bits, Enum, Ord, Real, Integral, Bounded, NFData)
 
 instance KnownSymbol (ToChTypeName (ChDateTime tz)) => IsChType (ChDateTime tz)
   where
@@ -2052,7 +2051,7 @@ instance FromChType (ChDateTime tz) UTCTime    where fromChType (MkChDateTime w3
 
 
 newtype ChDate = MkChDate Word16
-  deriving newtype (Show, Eq, Prim, Bits, Bounded, Enum, NFData)
+  deriving newtype (Show, Eq, Bits, Bounded, Enum, NFData)
 
 instance IsChType ChDate where
   type ToChTypeName ChDate = "Date"
@@ -2110,7 +2109,7 @@ instance
   Part of protocol implementation
 -}
 newtype UVarInt = MkUVarInt Word64
-  deriving newtype (Show, Eq, Num, Prim, Bits, Enum, Ord, Real, Integral, Bounded, NFData)
+  deriving newtype (Show, Eq, Num, Bits, Enum, Ord, Real, Integral, Bounded, NFData)