+++ ChangeLog.md
@@ -9,6 +9,7 @@ ClickHaskell documentation got it's own domain name: https://clickhaskell.dev/
 ## Features:
 - Query serialization support for UUID (+ add test)
 - Export of client errors for exception handling
+- Dropped vector dependency
 
 ## API changes:
 - ### DateTime type now parametrized with timezone